+++ docs/lib/snippets/isolates.dart
@@ -50,8 +50,8 @@ void main() async {
   // internally. This is NOT what we returned from _backgroundConnection, drift
   // uses an internal proxy class for isolate communication.
   // As long as the isolate is used by only one database (it is here), we can
-  // use `shutdownOnClose` to dispose the isolate after closing the connection.
-  final connection = await isolate.connect(shutdownOnClose: true);
+  // use `singleClientMode` to dispose the isolate after closing the connection.
+  final connection = await isolate.connect(singleClientMode: true);
 
   final db = TodoDb.connect(connection);
 
@@ -69,7 +69,7 @@ void connectSynchronously() {
   TodoDb.connect(
     DatabaseConnection.delayed(Future.sync(() async {
       final isolate = await DriftIsolate.spawn(_backgroundConnection);
-      return isolate.connect(shutdownOnClose: true);
+      return isolate.connect(singleClientMode: true);
     })),
   );
   // #enddocregion delayed
@@ -123,7 +123,7 @@ class _IsolateStartRequest {
 DatabaseConnection createDriftIsolateAndConnect() {
   return DatabaseConnection.delayed(Future.sync(() async {
     final isolate = await _createDriftIsolate();
-    return await isolate.connect(shutdownOnClose: true);
+    return await isolate.connect(singleClientMode: true);
   }));
 }
 // #enddocregion init_connect